requirements.txt
pandas
feedparser
pyyaml
matplotlib
jupyter
transformers
torch
streamlit
python-dotenv
pytest
rapidfuzz
